feat: rename sparse ss
