+++ src/matrix/mod.rs
@@ -13,6 +13,7 @@ pub mod operators;
 pub mod po;
 pub mod pt;
 pub mod sp;
+pub mod ss;
 pub mod st;
 pub mod sy;
 pub mod to;
