+++ src/matrix/ss/mul.rs
