+++ src/matrix/ss/mul_lhs.rs
