+++ src/matrix/ss/mul_rhs.rs
